Infopoints refactoring (#439)
* Modularized the screen to infopoint map parsers

* Introduced InfopointStatusMap type

* Renamed variable 'infopointOpenStatusMap' to 'infopointStatusMap' since infopoint can carry also other type of information than just open status

* Extracted infopoint auto mobile closing functionality to its separate hook

* Extracted infopoint closing to its separate hook

* Moved infopoint components to newly created 'components' folder

* Divided one big component ScreenAnchorComponent to more smaller components

* Moved TooltipInfopoint and its components to separate folder so they are grouped together

* fix: infopoint model typing, marked fields which can be undefined as optional

* Reordered imports in TooltipInfopoint component

* Merged BasicTooltipInfopoint with TooltipInfopoint component

* fix: replaced useElementSize with useResizeObserver in all occurenses
