+++ web/src/components/infopoint/components/tooltip-infopoint/InfopointBody.tsx
