+++ web/src/components/infopoint/screen-to-map-parsers/image-parser.ts
@@ -1,0 +1,19 @@
+import { ImageScreen } from "models";
+import { InfopointStatusMap } from "../useTooltipInfopoint";
+
+export const parseImageScreenMap = (viewScreen: ImageScreen) => {
+  const infopointsMap = viewScreen?.infopoints?.reduce(
+    (acc, currInfopoint, currInfopointIndex) => {
+      return {
+        ...acc,
+        [`${currInfopointIndex}`]: {
+          isOpen: currInfopoint.alwaysVisible,
+          isAlwaysVisible: currInfopoint.alwaysVisible,
+        },
+      };
+    },
+    {}
+  );
+
+  return infopointsMap as InfopointStatusMap;
+};
