+++ web/src/components/infopoint/screen-to-map-parsers/slideshow-parser.ts
@@ -1,0 +1,26 @@
+import { SlideshowScreen } from "models";
+import { InfopointStatusMap } from "../useTooltipInfopoint";
+
+export const parseSlideshowScreenMap = (viewScreen: SlideshowScreen) => {
+  const infopointsMap = viewScreen.images?.reduce(
+    (acc, currImage, currImageIndex) => {
+      const reducedImage = currImage.infopoints.reduce(
+        (innerAcc, currInfopoint, currInfopointIndex) => {
+          return {
+            ...innerAcc,
+            [`${currImageIndex}-${currInfopointIndex}`]: {
+              isOpen: currInfopoint.alwaysVisible,
+              isAlwaysVisible: currInfopoint.alwaysVisible,
+            },
+          };
+        },
+        {}
+      );
+
+      return { ...acc, ...reducedImage };
+    },
+    {}
+  );
+
+  return infopointsMap as InfopointStatusMap;
+};
