+++ web/src/containers/views/view-image/view-image.tsx
@@ -9,7 +9,7 @@ import { useSelector } from "react-redux";
 import { createSelector } from "reselect";
 
 import { useSpring, animated } from "react-spring";
-import useElementSize from "hooks/element-size-hook";
+import useResizeObserver from "hooks/use-resize-observer";
 import useTooltipInfopoint from "components/infopoint/useTooltipInfopoint";
 import { useGlassMagnifier } from "hooks/view-hooks/glass-magnifier-hook/useGlassMagnifier";
 
@@ -36,7 +36,7 @@ export const ViewImage = ({ screenPreloadedFiles }: ScreenProps) => {
 
   /* Wrapper is the whole <div> of this screen */
   /* Container is <div> without tooltip */
-  const [wrapperRef, parentSize] = useElementSize();
+  const [wrapperRef, parentSize] = useResizeObserver();
   const containerRef = useRef<HTMLDivElement>(null);
   const [containedImgEl, setContainedImgEl] = useState<HTMLImageElement | null>(
     null
@@ -94,10 +94,10 @@ export const ViewImage = ({ screenPreloadedFiles }: ScreenProps) => {
   // - - - -
 
   const {
-    infopointOpenStatusMap,
-    setInfopointOpenStatusMap,
+    infopointStatusMap,
+    setInfopointStatusMap,
     closeInfopoints,
-    ScreenAnchorInfopoint,
+    AnchorInfopoint,
     TooltipInfoPoint,
   } = useTooltipInfopoint(viewScreen);
 
@@ -186,7 +186,7 @@ export const ViewImage = ({ screenPreloadedFiles }: ScreenProps) => {
 
             return (
               <React.Fragment key={`infopoint-tooltip-${infopointIndex}`}>
-                <ScreenAnchorInfopoint
+                <AnchorInfopoint
                   id={`infopoint-tooltip-${infopointIndex}`}
                   left={adjustedLeft}
                   top={adjustedTop}
@@ -202,8 +202,8 @@ export const ViewImage = ({ screenPreloadedFiles }: ScreenProps) => {
                   key={`infopoint-tooltip-${infopointIndex}`}
                   id={`infopoint-tooltip-${infopointIndex}`}
                   infopoint={infopoint}
-                  infopointOpenStatusMap={infopointOpenStatusMap}
-                  setInfopointOpenStatusMap={setInfopointOpenStatusMap}
+                  infopointStatusMap={infopointStatusMap}
+                  setInfopointStatusMap={setInfopointStatusMap}
                   primaryKey={infopointIndex.toString()}
                 />
               </React.Fragment>
